created TermGroup.pm with test file
TermGroup.pm represents a single <termGroup> element. It contains only
simple scalars, so I have used Class::Accessor to generate all of the
functionality. Each accessor and the constructor are documented, and
other POD such as the SYNOPSIS have been added as necessary.

The test file checks the constructor and all accessors for correctness.