created LangGroup.pm with test file
LangGroup.pm represents a single <langGroup> element of TBX-Min. It
contains a single scalar (its language), along with an accessor for
the contained term groups.

I chose to return the internal array reference for the term groups
because it seemed like it might be unwieldy to also provide
remove_term_group and clear_term_groups.

The term_groups accessor croaks on input so as to avoid confusion with
the getter/setter accessors in TermGroup.pm.