created ConceptEntry.pm with test file
ConceptEntry.pm represents a single <conceptEntry> element of TBX-Min.
It contains a single scalar (its id), along with an accessor for
the contained language groups.

I chose to return the internal array reference for the language groups
because it seemed like it might be unwieldy to also provide
remove_lang_group and clear_lang_groups.

The lang_groups accessor croaks on input so as to avoid confusion with
the getter/setter accessors in TermGroup.pm.

The test file checks the constructor and both accessors.