fixed id check in ConceptEntry test
The test for the id method was not actually checking much, since
$args->{id} did not exist. I suppose it was checking that setting the
id to undef leaves it undef, but since that is the default value it
doesn't help much.

Now id is set to a real string, and the return value of the id method
is checked and found not to be undef.