moved subject_field from document into concept
Because the subject_field is normally contained at the concept level in
other dialects of TBX, it has been moved into the concept of TBX-Min.
This required updating the sample file, and then updating Min.pm and
ConceptEntry.pm to reflect the change in structure, and then updating
their tests.