pre-release cleanups
Cleanup the code so that dzil test --release passes. Fix synopses so
that they compile, and satisfy perlcritic where possible. Also remove
$node from _termGrpStart, since it is an unused variable.

I also removed the signature line from dist.ini because it seemed
inappropriate to sign a TRG distribution with a private key. As far as
I know, we do not have a group key, and we do not sign things via
gpg4win.