rename "origin" to "creator"
Rename the "origin" field to "creator" to make the field's purpose
more clear. As the "origin" field is the equivalent of the UTX
"creator" field, it should hold the name of the person or organization
that created the glossary. "origin" was a confusing name, since it
could mean the source dictionary title or something similar.