rename "title" to "id"
Rename the "title" attribute to be "id" instead.
The document "title" was meant to be equivalent to the UTX dictionary
ID. This should be a string uniquely identifying the document.
Converting the file ID to "title" seemed sub-par, as titles and IDs are
very different. Although a title was nice to have, the ID can be used
as a title if needed (and a file description will be added in the
future).