give picklist values a TBX bias
TBX-Min is to help UTX users migrate to TBX, not the other way around.
Instead of having a neutral or UTX bias in the partOfSpeech and
termStatus picklists, take the values from the TBX standard.

This means that conversion of a UTX "sentence" part of speech will become
"other", "provisional" term status will become "admitted", etc.